+++ NEWS.md
@@ -1,6 +1,10 @@
+# sircovid 0.2.19
+
+- Fix name error with p_death_hosp/_D in new odin model
+
 # sircovid 0.2.18
 
-- Fix bug with p_death_hosp/_D in new odin model
+- Fix dates used in pMCMC
 
 # sircovid 0.2.17